% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_lab_generator.R
\name{adm_lab_generator}
\alias{adm_lab_generator}
\title{Admission and lab merged TB generator}
\usage{
adm_lab_generator(adm, lab, baseline_days = 90)
}
\arguments{
\item{adm}{data frame containing TB admission data cleaned and allocated object
class by tbcleanr package}

\item{lab}{data frame containing TB laboratory data cleaned and allocated object
class by tbcleanr package}

\item{baseline_days}{criteria for using pre-treatment samples}
}
\description{
Designed to generate additional variables from within the merged TB 
admission and laboratory data file. Baseline DST, culture conversion 
dates and baseline smear and culture are generated and outputed with 
the id number of the data set.
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
