% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_generator.R
\name{age_generator}
\alias{age_generator}
\title{Calculate age from two dates}
\usage{
age_generator(x, categorise = FALSE, paediatric = FALSE,
  rm_orig = TRUE)
}
\arguments{
\item{x}{data frame containing TB admission variables}

\item{categorise}{logical - generate additional factor age variable}

\item{paediatric}{logical - generate additional paediatric age factor variable}

\item{rm_orig}{remove original variables - TRUE or FALSE}
}
\description{
Use date of birth and date of starting treatment to calculate 
age at baseline. Takes data frame with date of birth and start 
of treatment date to generate age variable in years.
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar
}
