% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_dst.R
\name{aggregate_dst}
\alias{aggregate_dst}
\title{Aggregate DST}
\usage{
aggregate_dst(x, drug_class)
}
\arguments{
\item{x}{data frame of laboratory DST data}

\item{drug_class}{string used to define data frame variables to use
to calculate aggregate DST results}
}
\value{
data frame with one factorised column of DST results with 
possible results include "Sensitive" or "Resistant"
}
\description{
Generate laboratory sample DST results by prioritising resistant
results over sensitive results where more than one DST test has 
been performed
}
