% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{baseline}
\alias{baseline}
\title{Calculate baseline culture or smear status}
\usage{
baseline(x, project = c("kk", "chechnya"), test = c("culture", "smear"),
  days = 90)
}
\arguments{
\item{x}{data frame or data.table with combined admission and lab data. Must contain
patient ID number, start treatment date, sample date, and test result}

\item{project}{define project location to apply.
Values can be "kk", "chechnya".}

\item{test}{define which baseline test result to check}

\item{days}{criteria for using pre-treatment samples}
}
\description{
This function takes a data frame and uses a unique patient identifier, dates of treatment start and sample submission and sample results to provide a baseline culture or smear result.
}
\seealso{
\code{\link{tbgeneratr}}
}
\keyword{TB}
