% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseliner.default.R, R/drug_timer.default.R
\name{baseliner.default}
\alias{baseliner.default}
\title{Default method for baseliner()}
\usage{
\method{baseliner}{default}(adm, change, drug)

\method{baseliner}{default}(adm, change, drug)
}
\arguments{
\item{adm}{admission data set cleaned using `tbcleanr`}

\item{change}{change data set cleaned using `tbcleanr`}

\item{drug}{define which drug to calculate}

\item{x}{data frame containing variables}
}
\description{
Allow data frames with unspecified object class to pass through

Allow data frames with unspecified object class to pass through
}
\seealso{
\code{\link{tbgeneratr}}

\code{\link{tbgeneratr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}

Jay Achar \email{jay.achar@doctors.org.uk}
}
