% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseliner.epiinfo.R
\name{baseliner.epiinfo}
\alias{baseliner.epiinfo}
\title{Baseline culture or smear status}
\usage{
\method{baseliner}{epiinfo}(adm, lab, baseline_test = c("culture",
  "smear"), baseline_days = 90)
}
\arguments{
\item{adm}{data frame containing TB admission data cleaned and allocated object
class by tbcleanr package}

\item{lab}{data frame containing TB laboratory data cleaned and allocated object
class by tbcleanr package}

\item{baseline_test}{string to define which baseline test result to check}

\item{baseline_days}{number of days prior to treatment start to define baseline 
acceptable period}
}
\description{
This function takes two data frames, including TB admission and 
laboratory programme data, uses a unique patient identifier, dates of 
treatment start and sample submission and sample results to calculate baseline 
culture or smear status. The function requires admission and laboratory data to 
have been prepared with the tbcleanr package and for the admission data to have 
class of either "epiinfo" or "koch6"
}
\seealso{
\code{\link{tbgeneratr}}
}
\keyword{TB}
