% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi_generator.default.R
\name{bmi_generator.default}
\alias{bmi_generator.default}
\title{Default method for bmi_generator()}
\usage{
\method{bmi_generator}{default}(x)
}
\arguments{
\item{x}{data frame containing variables}
}
\description{
Allow data frames with unspecified object class to pass through
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
