% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi_generator.R
\name{bmi_generator}
\alias{bmi_generator}
\title{Calculate BMI from height and weight}
\usage{
bmi_generator(x, rm_orig = TRUE, ...)
}
\arguments{
\item{x}{data frame containing height and weight variables}

\item{rm_orig}{remove original variables - TRUE or FALSE}

\item{...}{further arguments passed to or from other methods}
}
\description{
Use numerical height and weight variables to calculate BMI
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
