% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\name{converter.default}
\alias{converter.default}
\title{Culture/smear conversion calculator}
\usage{
\method{converter}{default}(adm)
}
\arguments{
\item{adm}{data frame containing TB admission data cleaned and allocated object
class by tbcleanr package}
}
\description{
Takes laboratory data with start and end treatment informaiton and
calculates culture or smear conversion dates based on >30 days criteria
}
\seealso{
\code{\link{tbgeneratr}}
}
