% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\name{converter}
\alias{converter}
\title{Culture/smear conversion calculator}
\usage{
converter(x, convert_type = c("culture", "smear"), software = c("excel",
  "koch_6", "epiinfo"), project = c("kk", "chechnya"), file = c("adm",
  "lab", "clinical_lab"))
}
\arguments{
\item{x}{data frame}

\item{convert_type}{define what type of culture conversion required.}

\item{software}{define software used for data collection.
Values can be "excel", "koch_6", "epiinfo"}

\item{project}{define project location to apply.
Values can be "kk", "chechnya".}

\item{file}{define database file argument to apply.
Values can be "adm", "lab", "clinical_lab",}

\item{...}{further arguments passed to or from other methods}
}
\description{
Takes laboratory data with start and end treatment informaiton and
calculates culture or smear conversion dates based on >30 days criteria
}
\examples{
\dontrun{
converter(p, convert_type = "culture", software = "koch_6", project = "chechnya", file = "adm")
}
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
