% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_baseliner.R
\name{drug_baseliner}
\alias{drug_baseliner}
\title{Baseline drug DST}
\usage{
drug_baseliner(x, drug, days = 30)
}
\arguments{
\item{x}{data frame from dst_baseliner()}

\item{drug}{name of drug dst variable from dst_baseliner()}

\item{days}{number of days allowed between individual drug specimen 
and baseline specimen}
}
\description{
Find baseline drug DST after baseline rifampicin result defined - 
auxillary function for use in dst_baseliner()
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar
}
