% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_timer.epiinfo.R
\name{drug_timer.epiinfo}
\alias{drug_timer.epiinfo}
\title{Drug duration calculator}
\usage{
\method{drug_timer}{epiinfo}(adm, change, drug)
}
\arguments{
\item{adm}{admission data set cleaned using `tbcleanr`}

\item{change}{change data set cleaned using `tbcleanr`}

\item{drug}{define which drug to calculate}
}
\value{
admission data frame with additional drug days variable added
}
\description{
This function calculates the number of days a specific drug was prescribed
to a patient. It uses information from the admission data set to establish 
the baseline treatment regimen, then adds days based on treatment adjustments
documented in the change data set.
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar
}
