% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_baseliner.default.R
\name{dst_baseliner.default}
\alias{dst_baseliner.default}
\title{Default method for dst_baseliner()}
\usage{
\method{dst_baseliner}{default}(adm, lab, dst_time = 90, dst_days = 30)
}
\arguments{
\item{adm}{data frame containing TB admission data cleaned and allocated object
class by tbcleanr package}

\item{lab}{data frame containing TB laboratory data cleaned and allocated object
class by tbcleanr package}

\item{dst_time}{absolute historical limit for including specimens}

\item{dst_days}{additional criteria for including non-rif results}
}
\description{
Allow data frames with unspecified object class to pass through
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
