% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_baseliner.epiinfo.R
\name{dst_baseliner.epiinfo}
\alias{dst_baseliner.epiinfo}
\title{Baseline DST calculator}
\usage{
\method{dst_baseliner}{epiinfo}(adm, lab, dst_time = 90, dst_days = 30)
}
\arguments{
\item{adm}{data frame containing TB admission data cleaned and allocated object
class by tbcleanr package}

\item{lab}{data frame containing TB laboratory data cleaned and allocated object
class by tbcleanr package}

\item{dst_time}{absolute historical limit for including specimens}

\item{dst_days}{additional criteria for including non-rif results}
}
\description{
Takes a TB DST laboratory dataset with information on 
treatment start time to output the baseline DST according
to specific rules designed to reduce missingness
}
\seealso{
\code{\link{tbgeneratr}}
}
