% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_baseliner.R
\name{dst_baseliner}
\alias{dst_baseliner}
\title{DST baseliner}
\usage{
dst_baseliner(x, project = c("kk", "chechnya"), time = 90, days = 30)
}
\arguments{
\item{x}{data frame containing all DST data and start/end dates}

\item{project}{define which project's data is being used}

\item{time}{define pre-treatment period to use specimens}

\item{days}{define how many days to use to recover other drug DST}
}
\description{
Generate a baseline DST for each subject using a
combination of laboratory DST data and treatment start 
and end dates. Detects rifampicin result closest to 
treatment start, then uses specimens with same rifampicin
result within 'days' of this specimen to generate other drug
results. Only specimens up to 'time' before treatment start 
and < 7 days after treatment start are included
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
