% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_amplify.R, R/fail_ltfu.R
\name{fail_amplify.default}
\alias{fail_amplify.default}
\title{Fail - culture reversion}
\usage{
\method{fail_amplify}{default}(adm, adhere)

\method{fail_amplify}{default}(adm, adhere)
}
\arguments{
\item{adm}{data frame of individual patient admission records cleaned 
with `tbcleanr` and combined with laboratory data using \code{tbgeneratr::adm_lab_generator()}. 
Must also include baseline drug DST generated data from \code{tbgeneratr::dst_baseliner()}}

\item{adhere}{data frame of individual patient follow-up data including
monthly drug-specific adherence data cleaned by `tbcleanr::adhere_cleanr()`}
}
\description{
Fail - culture reversion

Fail - LTFU
}
\seealso{
\code{\link{tbgeneratr}}

\code{\link{tbgeneratr}}
}
\author{
Jay Achar

Jay Achar
}
