% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_amplify.R
\name{fail_amplify}
\alias{fail_amplify}
\title{Fail - amplify resistance}
\usage{
fail_amplify(adm, lab)
}
\arguments{
\item{adm}{data frame of individual patient admission records cleaned 
with `tbcleanr` and combined with laboratory data using \code{tbgeneratr::adm_lab_generator()}. 
Must also include baseline drug DST generated data from \code{tbgeneratr::dst_baseliner()}}

\item{lab}{data frame of laboratory TB data cleaned using `tbcleanr`}
}
\description{
This function detects whether a record diagnosed with DR-TB
should receive a failure outcome due to amplification of resistance.
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar
}
