% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_culture_reversion.R
\name{fail_culture_reversion.default}
\alias{fail_culture_reversion.default}
\title{Fail - culture reversion}
\usage{
\method{fail_culture_reversion}{default}(adm, lab)
}
\arguments{
\item{adm}{data frame of individual patient admission records cleaned 
with `tbcleanr` and combined with laboratory data using `tbgeneratr::adm_lab_generator()`}

\item{lab}{data frame of laboratory TB data cleaned using `tbcleanr`}
}
\description{
Fail - culture reversion
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar
}
