% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_culture_reversion.epiinfo.R
\name{fail_culture_reversion.epiinfo}
\alias{fail_culture_reversion.epiinfo}
\title{Fail - culture reversion}
\usage{
\method{fail_culture_reversion}{epiinfo}(adm, lab)
}
\arguments{
\item{adm}{data frame of individual patient admission records cleaned 
with `tbcleanr` and combined with laboratory data using `tbgeneratr::adm_lab_generator()`}

\item{lab}{data frame of laboratory TB data cleaned using `tbcleanr`}
}
\description{
This function detects whether a record diagnosed with DR-TB
should receive a failure outcome due to culture reversion
during the continuation phase. The WHO 2014 definitions suggests this
time-point to be 8 months after treatment commencement.
}
\seealso{
\code{\link{tbgeneratr}}
}
