% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_drugs.epiinfo.R
\name{fail_drugs.epiinfo}
\alias{fail_drugs.epiinfo}
\title{Fail - 2 drugs changed}
\usage{
\method{fail_drugs}{epiinfo}(adm, change, stop_days = 30)
}
\arguments{
\item{adm}{data frame of individual patient admission records cleaned 
with `tbcleanr`.}

\item{change}{data frame of individual patient follow-up data including
monthly drug-specific adherence data cleaned by `tbcleanr::adhere_cleanr()`}

\item{stop_days}{define interval in days where two drugs being stopped would result in failure.}
}
\description{
This function detects whether a record diagnosed with DR-TB
should receive a failure outcome due to two drugs being ceased within a designated interval.
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar
}
