% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_ltfu.R
\name{fail_ltfu.default}
\alias{fail_ltfu.default}
\title{Fail - LTFU}
\usage{
\method{fail_ltfu}{default}(adm, adhere)
}
\arguments{
\item{adm}{data frame of individual patient admission records cleaned 
with `tbcleanr`.}

\item{adhere}{data frame of individual patient follow-up data including
monthly drug-specific adherence data cleaned by `tbcleanr::adhere_cleanr()`}
}
\description{
Fail - LTFU
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar
}
