% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_ltfu.epiinfo.R
\name{fail_ltfu.epiinfo}
\alias{fail_ltfu.epiinfo}
\title{Fail - LTFU}
\usage{
\method{fail_ltfu}{epiinfo}(adm, adhere)
}
\arguments{
\item{adm}{data frame of individual patient admission records cleaned 
with `tbcleanr`.}

\item{adhere}{data frame of individual patient follow-up data including
monthly drug-specific adherence data cleaned by `tbcleanr::adhere_cleanr()`}
}
\description{
This function detects whether a record diagnosed with DR-TB
should receive a failure outcome due to fulfilling the criteria for loss to follow-up.
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar
}
