% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_no_culture_conversion.R
\name{fail_no_culture_conversion.default}
\alias{fail_no_culture_conversion.default}
\title{Fail - no culture conversion}
\usage{
\method{fail_no_culture_conversion}{default}(adm_lab)
}
\arguments{
\item{adm_lab}{data frame of individual patient records cleaned 
with `tbcleanr` and combined with laboratory data using `tbgeneratr::adm_lab_generator()`}
}
\description{
This function detects with a record diagnosed with DR-TB
should receive a failure outcome due to lack of culture conversion
before a defined time-point. The WHO 2014 definitions suggests this
time-point to be 8 months after treatment commencement.
}
\seealso{
\code{\link{tbgeneratr}}
}
