% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_no_culture_conversion.R
\name{fail_no_culture_conversion}
\alias{fail_no_culture_conversion}
\title{Fail - no culture conversion}
\usage{
fail_no_culture_conversion(adm_lab, no_cc_days = 240)
}
\arguments{
\item{adm_lab}{data frame of individual patient records cleaned 
with `tbcleanr` and combined with laboratory data using `tbgeneratr::adm_lab_generator()`}

\item{no_cc_days}{time cut-off in days to define failure to culture convert. 
Default value (240 days) taken from 2014 WHO definition of 8 months}
}
\description{
This function detects with a record diagnosed with DR-TB
should receive a failure outcome due to lack of culture conversion
before a defined time-point. The WHO 2014 definitions suggests this
time-point to be 8 months after treatment commencement.
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar
}
