% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_outcome.epiinfo.R
\name{finalise_outcome.epiinfo}
\alias{finalise_outcome.epiinfo}
\title{Finalise TB treatment outcome}
\usage{
\method{finalise_outcome}{epiinfo}(x, bin_outcome = TRUE, ...)
}
\arguments{
\item{x}{data frame containing outcome variables}

\item{bin_outcome}{create new binary factor variable simplifying treatment outcome}

\item{...}{further arguments passed to or from other methods}
}
\description{
Finalise TB outcome variable to align with WHO definitions
}
\details{
This is a generic function. Methods for EpiInfo and Koch6 data have been defined 
within this package. They are designed to standardise outcome factor levels
according to \href{https://www.who.int/tb/publications/definitions/en/}{WHO (2013) TB definitions}.

If \code{bin_outcome} is \code{TRUE}, an additional simplified factor outcome variable is generated with 
levels \code{Successful}, \code{Unsuccessful} and \code{Not evaluated}, corresponding to definitions
in the WHO document.

The original data frame with generated variables appended is returned.

Any original data not represented in the WHO definitions are converted to \code{NA}.
}
\seealso{
\code{\link{tbgeneratr}}
}
