% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassign_lab_no.R
\name{reassign_lab_no.default}
\alias{reassign_lab_no.default}
\title{Reassign lab numbers}
\usage{
\method{reassign_lab_no}{default}(adm, lab)
}
\arguments{
\item{adm}{data frame of admission data cleaned with the tbcleanr package}

\item{lab}{data frame of laboratory data cleaned with the tbcleanr package}
}
\value{
data frame containing lab data with updated patient ID numbers where 
they were previously missing
}
\description{
Check admission data set for diagnostic laboratory sample 
identification numbers and link to laboratory data set where
this is absent in original data.
}
\seealso{
\code{\link{tbgeneratr}}
}
