% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalculate_outcome.R
\name{recalculate_outcome.default}
\alias{recalculate_outcome.default}
\title{Recalculate outcome}
\usage{
\method{recalculate_outcome}{default}(adm, adhere, change, lab,
  stop_days = 30)
}
\arguments{
\item{adm}{data frame of individual patient admission records cleaned 
with `tbcleanr`.}

\item{adhere}{data frame of individual patient follow-up data including
monthly drug-specific adherence data cleaned by `tbcleanr::adhere_cleanr()`}

\item{change}{data frame of individual patient follow-up data including
monthly drug-specific change data cleaned by `tbcleanr::change_cleanr()`}

\item{lab}{data frame of laboratory TB data cleaned using `tbcleanr`}

\item{stop_days}{define interval in days where two drugs being stopped would result in failure.}
}
\description{
Recalculate outcome
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar
}
