% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regimen_calculator.epiinfo.R
\name{regimen_calculator.epiinfo}
\alias{regimen_calculator.epiinfo}
\title{Regimen calculator}
\usage{
\method{regimen_calculator}{epiinfo}(adm, change, days = 0L)
}
\arguments{
\item{adm}{data frame of individual patient admission records cleaned 
with `tbcleanr`.}

\item{change}{data frame of TB  drug change data cleaned using `tbcleanr`}

\item{days}{single integer to define the timepoint following treatment initiation 
to generate regimen string}
}
\description{
This function will generate a string indicating the treatment regimen at any given 
date following initiation based on baseline regimen and drug change data.
}
\seealso{
\code{\link{tbgeneratr}}
}
