% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_splittr.R
\name{start_splittr}
\alias{start_splittr}
\title{Split start treatment variable into year/month/day}
\usage{
start_splittr(x, rm_orig = FALSE)
}
\arguments{
\item{x}{data frame containing cleaned TB admission data}

\item{rm_orig}{remove original variables - TRUE or FALSE}
}
\description{
Takes cleaned admission data frame and splits start treatment date variable
into component parts (year, month, day) to allow filtered analyses.
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
