% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbgeneratr.R
\docType{package}
\name{tbgeneratr}
\alias{tbgeneratr}
\alias{tbgeneratr-package}
\title{tbgeneratr: A package for generating TB data.}
\description{
The tbgeneratr package provides functions to generate TB data  
from clean data sets.
}
\seealso{
\code{\link{age_generator}}

\code{\link{baseline}}

\code{\link{bmi_generator}}

\code{\link{tb.convert}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
