% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_generator.R
\name{adm_generator}
\alias{adm_generator}
\title{Admission TB variable generator}
\usage{
adm_generator(x, rm_orig = TRUE, categorise = TRUE,
  paediatric = TRUE, bin_outcome = TRUE, ...)
}
\arguments{
\item{x}{data frame of cleaned TB admission data using the `tbcleanr` package}

\item{rm_orig}{logical argument to remove original variables from age and BMI
calculators}

\item{categorise}{logical flag to add factor variable of ages}

\item{paediatric}{logical flag to add factor variable for childhood ages}

\item{bin_outcome}{logical flag to add binary outcome variable}

\item{...}{additional arguments to pass}
}
\description{
Designed to generate additional variables from within the TB admission
data file. Age and BMI are calculated from date of birth, start treatment
date, height and weight. Start treatment date is also split into components
to allow easier grouping by year, month or date.
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar
}
