% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_generator.R
\name{adm_generator}
\alias{adm_generator}
\title{Admission TB generator}
\usage{
adm_generator(x, rm_orig = TRUE, ...)
}
\arguments{
\item{x}{data frame of cleaned TB admission data}

\item{rm_orig}{logical argument to remove original variables}

\item{...}{logical arguments for \code{age_generator()}:
\itemize{
  \item \code{categorise} to add factor variable of ages,
  \item \code{paediatric} to add factor variable for childhood ages.
}}
}
\description{
Designed to generate additional variables from within the TB admission
data file. Age and BMI are calculated from date of birth, start treatment
date, height and weight. Start treatment date is also split into components
to allow easier grouping by year, month or date.
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar
}
