% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_lab_generator.R
\name{adm_lab_generator}
\alias{adm_lab_generator}
\title{Admission and lab merged TB generator}
\usage{
adm_lab_generator(x, software = c("excel", "koch_6", "epiinfo"),
  project = c("kk", "chechnya"), file = c("adm", "lab",
  "clinical_lab"), dst_time = 90, dst_days = 30, baseline_days = 90)
}
\arguments{
\item{x}{data frame of cleaned and merged TB admission and lab data}

\item{software}{define software used for data collection.
Values can be "excel", "koch_6", "epiinfo"}

\item{project}{define project location to apply.
Values can be "kk", "chechnya".}

\item{file}{define database file argument to apply.
Values can be "adm", "lab", "clinical_lab",}

\item{dst_time}{absolute historical limit for including specimens}

\item{dst_days}{additional criteria for including non-rif results}

\item{baseline_days}{criteria for using pre-treatment samples}
}
\description{
Designed to generate additional variables from within the merged TB 
admission and laboratory data file. Baseline DST, culture conversion 
dates and baseline smear and culture are generated and outputed with 
the id number of the data set.
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
