% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_generator.default.R
\name{age_generator.default}
\alias{age_generator.default}
\title{Default method for age_generator()}
\usage{
\method{age_generator}{default}(x, categorise = FALSE,
  paediatric = FALSE, rm_orig = TRUE)
}
\arguments{
\item{x}{data frame containing variables}

\item{categorise}{logical - generate additional factor age variable}

\item{paediatric}{logical - generate additional paediatric age factor variable}

\item{rm_orig}{remove original variables - TRUE or FALSE}
}
\description{
Allow data frames with unspecified object class to pass through
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
