% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_generator.R
\name{age_generator}
\alias{age_generator}
\title{Calculate age from two dates}
\usage{
age_generator(x, software = c("epiinfo", "koch_6", "excel"),
  rm_orig = TRUE)
}
\arguments{
\item{x}{data frame containing Koch 6 admission variables}

\item{software}{define software used for data collection.
Values can be "excel", "koch_6", "epiinfo"}

\item{rm_orig}{remove original variables - TRUE or FALSE}
}
\description{
Use date of birth and date of starting treatment to calculate 
age at baseline. Takes data frame with date of birth and start 
of treatment date to generate age variable in years.
}
\examples{
\dontrun{
age_generator(p, software = "koch_6", rm_orig = FALSE)
}
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
