% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseliner.default.R
\name{baseliner.default}
\alias{baseliner.default}
\title{Default method for baseliner()}
\usage{
\method{baseliner}{default}(x)
}
\arguments{
\item{x}{data frame containing variables}
}
\description{
Allow data frames with unspecified object class to pass through
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar
}
