% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseliner.R
\name{baseliner.default}
\alias{baseliner.default}
\title{Calculate baseline culture or smear status}
\usage{
\method{baseliner}{default}(adm)
}
\arguments{
\item{adm}{data frame containing TB admission data cleaned and allocated object
class by tbcleanr package}
}
\description{
This function takes two data frames, including TB admission and 
laboratory programme data, uses a unique patient identifier, dates of 
treatment start and sample submission and sample results to calculate baseline 
culture or smear status. The function requires admission and laboratory data to 
have been prepared with the tbcleanr package and for the admission data to have 
class of either "epiinfo" or "koch6"
}
\seealso{
\code{\link{tbgeneratr}}
}
