% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseliner.R
\name{baseliner}
\alias{baseliner}
\title{Calculate baseline culture or smear status}
\usage{
baseliner(x, software = c("koch_6", "epiinfo", "excel"),
  project = c("kk", "chechnya"), file = c("adm", "lab"),
  baseline_test = c("culture", "smear"), baseline_days = 90)
}
\arguments{
\item{x}{data frame or data.table with combined admission and lab data. Must contain
patient ID number, start treatment date, sample date, and test result}

\item{software}{define software used for data collection.
Values can be "excel", "koch_6", "epiinfo"}

\item{project}{define project location to apply.
Values can be "kk", "chechnya".}

\item{file}{define database file argument to apply.
Values can be "adm", "lab",}

\item{baseline_test}{define which baseline test result to check}

\item{baseline_days}{criteria for using pre-treatment samples}
}
\description{
This function takes a data frame and uses a unique patient identifier, dates of treatment start and sample submission and sample results to provide a baseline culture or smear result.
}
\seealso{
\code{\link{tbgeneratr}}
}
\keyword{TB}
