% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi_generator.R
\name{bmi_generator.default}
\alias{bmi_generator.default}
\title{Calculate BMI from height and weight}
\usage{
\method{bmi_generator}{default}(x)
}
\arguments{
\item{x}{data frame containing height and weight variables}
}
\description{
Use numerical height and weight variables to calculate BMI
}
\seealso{
\code{\link{tbgeneratr}}
}
