% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.default.R
\name{converter.default}
\alias{converter.default}
\title{Default method for converter()}
\usage{
\method{converter}{default}(x)
}
\arguments{
\item{x}{data frame containing variables}
}
\description{
Allow data frames with unspecified object class to pass through
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
