% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\name{converter}
\alias{converter}
\title{Culture/smear conversion calculator}
\usage{
converter(adm, lab, convert_type = c("culture", "smear"))
}
\arguments{
\item{adm}{data frame containing TB admission data cleaned and allocated object
class by tbcleanr package}

\item{lab}{data frame containing TB laboratory data cleaned and allocated object
class by tbcleanr package}

\item{convert_type}{string to define which baseline test result to check}
}
\description{
Takes laboratory data with start and end treatment informaiton and
calculates culture or smear conversion dates based on >30 days criteria
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar
}
