% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_baseliner.R
\name{dst_baseliner}
\alias{dst_baseliner}
\title{Baseline DST calculator}
\usage{
dst_baseliner(x, project = c("kk", "chechnya"), dst_time = 90,
  dst_days = 30)
}
\arguments{
\item{x}{data frame with TB and treatmetn data}

\item{project}{define which project this dataset refers to}

\item{dst_time}{absolute historical limit for including specimens}

\item{dst_days}{additional criteria for including non-rif results}
}
\description{
Takes a TB DST laboratory dataset with information on 
treatment start time to output the baseline DST according
to specific rules designed to reduce missingness
}
\examples{
\dontrun{
dst_baseliner(lab, project = "kk")
}
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar \email{jay.achar@doctors.org.uk}
}
