% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fail_no_culture_conversion.R
\name{fail_no_culture_conversion.default}
\alias{fail_no_culture_conversion.default}
\title{Fail - no culture conversion}
\usage{
\method{fail_no_culture_conversion}{default}(adm_lab)
}
\arguments{
\item{adm_lab}{data frame of individual patient records cleaned 
with `tbcleanr` and combined with laboratory data using `tbgeneratr::adm_lab_generator()`}
}
\description{
Fail - no culture conversion
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar
}
