% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_outcome.R
\name{finalise_outcome.default}
\alias{finalise_outcome.default}
\title{Finalise TB treatment outcome}
\usage{
\method{finalise_outcome}{default}(x, bin_outcome = TRUE, ...)
}
\arguments{
\item{x}{data frame containing outcome variables}

\item{bin_outcome}{create new binary factor variable simplifying treatment outcome}

\item{...}{further arguments passed to or from other methods}
}
\description{
Finalise TB outcome variable to align with WHO definitions
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar
}
