% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalise_outcome.epiinfo.R
\name{finalise_outcome.epiinfo}
\alias{finalise_outcome.epiinfo}
\title{Finalise TB treatment outcome}
\usage{
\method{finalise_outcome}{epiinfo}(x, simplify = TRUE, ...)
}
\arguments{
\item{x}{data frame containing outcome variables}

\item{simplify}{logical to add binary outcome variable}

\item{...}{further arguments passed to or from other methods}
}
\description{
Finalise TB outcome variable to align with WHO definitions
}
\seealso{
\code{\link{tbgeneratr}}
}
\author{
Jay Achar
}
